Make geodata_impl.h to allow for testing support functions without introducing Eigen dependency broadly
