+++ cmake/ExternalEigen.cmake
@@ -9,11 +9,8 @@ set(EIGEN_OPTIONS ${PALACE_SUPERBUILD_DEFAULT_ARGS})
 list(APPEND EIGEN_OPTIONS
   "-DCMAKE_CXX_COMPILER=${CMAKE_CXX_COMPILER}"
   "-DCMAKE_CXX_FLAGS=${CMAKE_CXX_FLAGS}"
-  "-DEIGEN_BUILD_DOC=ON"
+  "-DEIGEN_BUILD_DOC=OFF"
   "-DBUILD_TESTING=OFF"
-  "-DEIGEN_TEST_FORTRAN=OFF"
-  "-DEIGEN_BUILD_BLAS=OFF"
-  "-DEIGEN_BUILD_LAPACK=OFF"
   "-DCMAKE_Fortran_COMPILER=" # Unneeded
 )
