+++ palace/utils/CMakeLists.txt
@@ -10,6 +10,7 @@ target_sources(${LIB_TARGET_NAME}
   ${CMAKE_CURRENT_SOURCE_DIR}/configfile.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/dorfler.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/geodata.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/geodata_impl.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/iodata.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/meshio.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/omp.cpp
